Changes to exmode in eventer + minor tweaks to GUI appearance
New behaviour of exmode 1:

The exmode feature tells eventer what to do in relation to calculation of interevent intervals after each exclusions zone, and for the first event in each wave. With mode = 1, IEIs are calculated for these events from the last event preceeding the exclusion zone under the assumption that no events occurred during the exclusion zone. When merging multiple waves, eventer also assumes there was no gap between waves. With mode = 2, the occurence of an event during these times is considered ambiguous and so is assign a value of NaN.